Refactor comWriteConfig function and fix file extension check
The comWriteConfig function in the common.c and common.h files underwent a renaming to ss_com_write_config and the return statement's logic was also modified for better readability. Moreover, an inaccurate file extension check in the world.c file was rectified by removing an unnecessary period character in the comparison string.