Made improvements to how postfx handles render targets
When fetching the postfx render target, it now always returns the postfx render target - but if postprocessing is disabled, we'll draw the current state of the viewport render target into the postfx render target (this is currently not working, but that's the idea).

This also, more importantly, fixes the issue where the postfx was not respecting the current render target.

Signed-off-by: hogsy <markelswo@gmail.com>