Refactored model deserialization and improved cooking route
This commit focuses on updating the model deserialization process and improving the model cooking route. The main changes include additions of mesh indexing, improvements in checks for normals, UVs, and colors in models, and modifications to ease addition of new loading formats. Mesh rendering logic was removed from the actor drawing function and should be addressed in future commits. This lays groundwork for further development of model-related functionality in the engine.

Signed-off-by: hogsy <markelswo@gmail.com>