Continued gradually making progress on the new model cooking support.
Hook up the actual cooking method for models.
Made a little bit of progress on actually implementing what we need to convert an obj to whatever, but running out of fuel.
Made a start on the SMD serialisation, though for now this just operates off PLMModel, so it could work for anything loaded via that API, heh.
Uh and disable the header bar for viewports again, it would be better if we can do this via a context menu or something.