Add Material Browser to Forge Editor
Introduced a Material Browser feature to the Forge Editor for easier material selection and preview. Updated the renderer to support image previews for materials and added necessary actions and UI components to integrate the browser. Also included a new function for linear lookup in ACM branches. Added several methods and classes to support material browser functionality.

Signed-off-by: Mark Sowden <hogsy@oldtimes-software.com>